+++ tests/MySqlData.Tests/MySqlDataSelectValueFixture.cs
@@ -19,5 +19,7 @@ namespace MySqlData.Tests
 		public string SelectNoRows => MySqlDatabase.SelectNoRows;
 
 		public IReadOnlyCollection<DbType> SupportedDbTypes => MySqlDatabase.SupportedDbTypes;
+
+		public Type NullValueExceptionType => MySqlDatabase.NullValueExceptionType;
 	}
 }